+++ Source/ACE.Server/Physics/PhysicsObj.cs
@@ -10,6 +10,7 @@ using ACE.Server.Physics.Collision;
 using ACE.Server.Physics.Combat;
 using ACE.Server.Physics.Common;
 using ACE.Server.Physics.Hooks;
+using ACE.Server.Physics.Managers;
 using ACE.Server.WorldObjects;
 
 using log4net;
@@ -502,7 +503,7 @@ namespace ACE.Server.Physics
 
         public PhysicsObj GetObjectA(uint objectID)
         {
-            return ObjectMaint.GetObjectA(objectID);
+            return ServerObjectManager.GetObjectA(objectID);
         }
 
         public float GetRadius()
@@ -1491,7 +1492,7 @@ namespace ACE.Server.Physics
         public void TurnToObject(uint objectID, MovementParameters movementParams)
         {
             if (ObjMaint == null) return;
-            var obj = ObjectMaint.GetObjectA(objectID);
+            var obj = ServerObjectManager.GetObjectA(objectID);
             if (obj == null) return;
             var parent = obj.Parent != null ? obj.Parent : obj;
 
@@ -3177,7 +3178,7 @@ namespace ACE.Server.Physics
 
             foreach (var objectID in CollisionTable.Keys)
             {
-                var obj = ObjectMaint.GetObjectA(objectID);
+                var obj = ServerObjectManager.GetObjectA(objectID);
                 if (obj != null)
                     report_object_collision(obj, TransientState.HasFlag(TransientStateFlags.Contact));
             }
@@ -3250,7 +3251,7 @@ namespace ACE.Server.Physics
         {
             if (ObjMaint != null)
             {
-                var collision = ObjectMaint.GetObjectA(objectID);
+                var collision = ServerObjectManager.GetObjectA(objectID);
                 if (collision != null)
                 {
                     if (!collision.State.HasFlag(PhysicsState.ReportCollisionsAsEnvironment))
@@ -3603,7 +3604,7 @@ namespace ACE.Server.Physics
             ObjID = guid;
             ID = guid.Full;
 
-            ObjectMaint.AddServerObject(this);
+            ServerObjectManager.AddServerObject(this);
         }
 
         /// <summary>
@@ -3828,7 +3829,7 @@ namespace ACE.Server.Physics
             MakePositionManager();
             if (ObjMaint == null) return;
 
-            var objectA = ObjectMaint.GetObjectA(objectID);
+            var objectA = ServerObjectManager.GetObjectA(objectID);
             if (objectA == null) return;
             if (objectA.Parent != null)
                 objectA = Parent;
